TestReachabilityFenceWithBuffer
